src/requirements.txt
functions-framework==3.7.0
google-cloud-bigquery
google-cloud-secret-manager==2.21.0
pandas==2.2.3
pyarrow==18.0.0
python-decouple==3.8
requests==2.32.3
Flask
